個人の投稿一覧表示機能

+++ db/migrate/20210812090844_create_posts.rb
@@ -2,7 +2,7 @@ class CreatePosts < ActiveRecord::Migration[6.0]
   def change
     create_table :posts do |t|
       t.text       :map_link, null: false
-      t.integer    :distance, null: false
+      t.float      :distance, null: false
       t.string     :course,   null: false
       t.string     :slope,    null: false
       t.string     :traffic,  null: false
